+++ uefi-test-runner/src/proto/usb/mod.rs
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+pub fn test() {
+    info!("Testing USB protocols");
+
+    io::test();
+}
+
+mod io;