+++ uefi/src/proto/device_path/build.rs
@@ -79,7 +79,7 @@ pub struct DevicePathBuilder<'a> {
 
 impl<'a> DevicePathBuilder<'a> {
     /// Create a builder backed by a statically-sized buffer.
-    pub fn with_buf(buf: &'a mut [MaybeUninit<u8>]) -> Self {
+    pub const fn with_buf(buf: &'a mut [MaybeUninit<u8>]) -> Self {
         Self {
             storage: BuilderStorage::Buf { buf, offset: 0 },
         }
@@ -248,7 +248,7 @@ mod tests {
     };
     use core::slice;
 
-    fn path_to_bytes(path: &DevicePath) -> &[u8] {
+    const fn path_to_bytes(path: &DevicePath) -> &[u8] {
         unsafe { slice::from_raw_parts(path.as_ffi_ptr().cast::<u8>(), size_of_val(path)) }
     }