+++ uefi/src/proto/media/file/info.rs
@@ -132,8 +132,7 @@ impl Display for FileInfoCreationError {
         match self {
             Self::InsufficientStorage(bytes) => write!(
                 f,
-                "provided buffer was too small. need at least {} bytes",
-                bytes
+                "provided buffer was too small. need at least {bytes} bytes"
             ),
         }
     }
@@ -427,8 +426,8 @@ impl FileProtocolInfo for FileSystemVolumeLabel {}
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::runtime::{Daylight, Time, TimeParams};
     use crate::CString16;
+    use crate::runtime::{Daylight, Time, TimeParams};
     use alloc::vec;
 
     fn validate_layout<T: InfoInternal + ?Sized>(info: &T, name: &[Char16]) {
@@ -436,7 +435,11 @@ mod tests {
         assert_eq!(align_of_val(info), T::alignment());
         // Check the hardcoded name slice offset.
         assert_eq!(
-            unsafe { (name.as_ptr() as *const u8).offset_from(info as *const _ as *const u8) },
+            unsafe {
+                name.as_ptr()
+                    .cast::<u8>()
+                    .offset_from(core::ptr::from_ref(info).cast::<u8>())
+            },
             T::name_offset() as isize
         );
     }