+++ uefi/src/proto/mod.rs
@@ -32,6 +32,7 @@ pub mod shell_params;
 pub mod shim;
 pub mod string;
 pub mod tcg;
+pub mod usb;
 
 mod boot_policy;