+++ uefi-raw/CHANGELOG.md
@@ -1,5 +1,9 @@
 # uefi-raw - [Unreleased]
 
+## Added
+- Added `AllocateType`.
+- Added `PciRootBridgeIoProtocol`.
+
 
 # uefi-raw - 0.11.0 (2025-05-04)
 
@@ -22,6 +26,11 @@
 ## Changed
 - `DevicePathProtocol` now derives
   `Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash`
+- The documentation for UEFI device paths has been streamlined and improved.
+
+## Changed
+
+- **Breaking:** The MSRV is now 1.85.1 and the crate uses the Rust 2024 edition.
 
 
 # uefi-raw - 0.10.0 (2025-02-07)