+++ uefi/CHANGELOG.md
@@ -2,13 +2,20 @@
 
 ## Added
 - Added `ConfigTableEntry::MEMORY_ATTRIBUTES_GUID` and `ConfigTableEntry::IMAGE_SECURITY_DATABASE_GUID`.
+- Added `proto::usb::io::UsbIo`.
 
 ## Changed
 - **Breaking:** `boot::stall` now take `core::time::Duration` instead of `usize`.
 - `table::cfg::*_GUID` constants now deprecated. Use `ConfigTableEntry::*_GUID` instead.
 - `system::with_config_table`, `system::with_stdin`, `system::with_stdout` and `system::with_stderr`
   now take mutably closure.
-
+- **Breaking:** The MSRV is now 1.85.1 and the crate uses the Rust 2024 edition.
+- The documentation in `lib.rs` now provides guidance on how to select features
+  tailored to your use case.
+- Feature `log-debugcon` is no longer a default feature. You only need to add
+  it in case you are also using the `logger` feature and if you run your UEFI
+  image in QEMU or Cloud Hypervisor, when the debugcon/debug-console device is
+  available.
 
 # uefi - 0.35.0 (2025-05-04)
 
@@ -56,6 +63,7 @@
   bugs on some devices.
 - The UEFI `allocator::Allocator` has been optimized for page-aligned
   allocations.
+- The documentation for UEFI device paths has been streamlined and improved.
 
 
 # uefi - 0.34.1 (2025-02-07)